% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-mixture_weights.R
\name{contradiction_plot}
\alias{contradiction_plot}
\title{Contradiction plot}
\usage{
contradiction_plot(
  miss.assumption,
  pi.1,
  varpi.11,
  varpi.10,
  lambda.0,
  bin.width,
  x.step
)
}
\arguments{
\item{miss.assumption}{Options are: "SNR" or "rER" (for stable noncomplier response / response exclusion restriction -- two names of the same assumption), or "SCR" (stable complier response)}

\item{pi.1, varpi.11, varpi.10, lambda.0}{Vectors for nuisance functions. If the missingness assumption is SNR/rER, provide \code{varpi.10}. If the missingness assumption is SCR, provide \code{varpi.11}.}
}
\description{
Make the diagnostic plot called "contradiction plot", which is a histogram of assumption-implied response probabilities under control, for compliers under the SNR/rER assumption and for noncompliers under the SCR assumption.
}
