% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-helpers.R
\name{.extract_dispersion_wtd_glm}
\alias{.extract_dispersion_wtd_glm}
\title{Extract the conditional distribution's dispersion param from a weighted GLM
(This function is borrowed from the package PIsens, almost verbatim.) Compute dispersion parameter from weighted models ignoring variance inflation due to weighting. Such variance inflation is desired when the dispersion parameter is used to compute SEs of regression coefficients. It is not desired when the dispersion parameter is used to estimate the conditional variance of the response variable, ie \code{var(Y|X)}.}
\usage{
.extract_dispersion_wtd_glm(mod)
}
\arguments{
\item{mod}{The fitted linear model object from \code{lm} or \code{glm}}
}
\description{
Extract the conditional distribution's dispersion param from a weighted GLM
(This function is borrowed from the package PIsens, almost verbatim.) Compute dispersion parameter from weighted models ignoring variance inflation due to weighting. Such variance inflation is desired when the dispersion parameter is used to compute SEs of regression coefficients. It is not desired when the dispersion parameter is used to estimate the conditional variance of the response variable, ie \code{var(Y|X)}.
}
\keyword{internal}
