% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-helpers.R
\name{.generalized_logit_approximate}
\alias{.generalized_logit_approximate}
\title{Approximate generalized logit transformation of a variable}
\usage{
.generalized_logit_approximate(x, bounds)
}
\arguments{
\item{x}{Variable to be transformed.}

\item{bounds}{A 2-length numeric vector with the lower and upper bounds of the variable.}
}
\description{
Approximate generalized logit transformation of a variable
}
\keyword{internal}
