% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b-EC_nuisance_estimation.R
\name{ec_nuisance_estimation}
\alias{ec_nuisance_estimation}
\title{Experience Corps specific: Estimate nuisance functions}
\usage{
ec_nuisance_estimation(dat)
}
\arguments{
\item{dat}{Data frame to estimate nuisance function on. Could include a variable \code{wt} for weighting units, which is useful for bootstrapping. If not including \code{wt}, units are equally weighted.}
}
\description{
Experience Corps specific: Estimate nuisance functions
}
