% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-effects_all_assumptions.R
\name{effects_all_assumptions}
\alias{effects_all_assumptions}
\title{Compute effects under all choices of assumptions
This function computes effects under all choices of assumptions as in Table 5 of the paper. The main use of this function is in bootstrapping.}
\usage{
effects_all_assumptions(nuis, sens.type, sens.params, y.bounds = NULL, weights)
}
\arguments{
\item{nuis}{The list of estimated nuisance functions. Like output of \code{ec_nuisance_estimation()}.}

\item{sens.type}{Type of sensitivity analysis (for PI violation) to be included.}

\item{sens.params}{Numeric vector including maximum and minimum values of the sensitivity parameter.}

\item{y.bounds}{Numeric vector holding the lower and upper bounds of the outcome. Needed only if \code{sens.type = "GOR"} and the outcome is not binary.}
}
\description{
Compute effects under all choices of assumptions
This function computes effects under all choices of assumptions as in Table 5 of the paper. The main use of this function is in bootstrapping.
}
