% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-effects_all_assumptions.R
\name{effects_all_assumptions_binary.y}
\alias{effects_all_assumptions_binary.y}
\title{Compute effects under all choices of assumptions in the case of a binary outcome
This function computes effects under all choices of assumptions as in Table 5 of the paper. The main use of this function is in bootstrapping. The key difference between this function and the function \code{effect_all_assumptions_binary.y} is that this one includes a "near ER" assumption, where "near" indicates an approximation similar to that in "near-SNR/SCR", but this is for the principal identification assumption ER itself. The motivation is that for a binary outcome, the ER assumption can imply mu01(X) values that are out of the \link{0,1} outcome bounds.}
\usage{
effects_all_assumptions_binary.y(
  nuis,
  sens.type,
  sens.params,
  y.bounds = NULL,
  weights,
  data,
  x.vars
)
}
\arguments{
\item{nuis}{The list of estimated nuisance functions. Like output of \code{j2_nuisance_estimation()} or \code{ec_nuisance_estimation()}.}

\item{sens.type}{Type of sensitivity analysis (for PI violation) to be included.}

\item{sens.params}{Numeric vector including maximum and minimum values of the sensitivity parameter.}

\item{y.bounds}{Numeric vector holding the lower and upper bounds of the outcome. Needed only if \code{sens.type = "GOR"} and the outcome is not binary.}
}
\description{
Compute effects under all choices of assumptions in the case of a binary outcome
This function computes effects under all choices of assumptions as in Table 5 of the paper. The main use of this function is in bootstrapping. The key difference between this function and the function \code{effect_all_assumptions_binary.y} is that this one includes a "near ER" assumption, where "near" indicates an approximation similar to that in "near-SNR/SCR", but this is for the principal identification assumption ER itself. The motivation is that for a binary outcome, the ER assumption can imply mu01(X) values that are out of the \link{0,1} outcome bounds.
}
