% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a-J2_nuisance_estimation.R
\name{j2_nuisance_estimation}
\alias{j2_nuisance_estimation}
\title{Jobs II specific: Estimate nuisance functions}
\usage{
j2_nuisance_estimation(dat)
}
\arguments{
\item{dat}{Data frame to estimate nuisance function on. Could include a variable \code{wt} for weighting units, which is useful for bootstrapping. If not including \code{wt}, units are equally weighted.}
}
\description{
Jobs II specific: Estimate nuisance functions
}
