% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-mixture_weights.R
\name{mixture_weights}
\alias{mixture_weights}
\title{Compute mixture weights}
\usage{
mixture_weights(miss.assumption, pi.1, varpi.11, varpi.10, lambda.0)
}
\arguments{
\item{miss.assumption}{A character string indicating the specific missingness assumption to be used. Options are: "nSNR" (near stable noncomplier response), "nSCR" (near stable complier response), "rPI" (principal ignorability for response) and "rPO" (proportional response odds)}

\item{pi.1, varpi.11, varpi.10, lambda.0}{Vectors for nuisance functions. Use the one(s) needed for the specific missingness assumption.}
}
\description{
Compute the mixture weights to be used in the outcome mixture equation based on the specific missingness assumption and estimated nuisance functions.
}
