% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-assemble_estimator.R
\name{plugin_estimate}
\alias{plugin_estimate}
\title{Compute the plug-in estimate of the CACE and NACE}
\usage{
plugin_estimate(pi.1, mu.11, mu.10, mu.01, mu.00)
}
\arguments{
\item{pi.1, mu.11, mu.10, mu.01, mu.00}{Vectors for estimated nuisance functions, computed at X values in the dataset}
}
\description{
Compute the plug-in estimate of the CACE and NACE
}
